requirements.txt
DateTime==4.3
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
Flask-WTF==0.14.3
gunicorn==20.0.4
matplotlib==3.3.3
pandas==1.2.0
